pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-blindspot_env"
version = "0.1.0"
description = "Blindspot — an OpenEnv environment for unknown-unknowns discovery & contextual onboarding."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "openai>=1.0.0",
    "websocket-client>=1.6.0",
    "numpy>=1.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]
precompute = [
    "requests>=2.31.0",
    "keybert>=0.8.0",
    "sentence-transformers>=2.7.0",
    "scikit-learn>=1.4.0",
    "feedparser>=6.0.0",
    "pandas>=2.2.0",
    "pyarrow>=16.0.0",
    "tqdm>=4.66.0",
    "openai>=1.50.0",
    "google-genai>=0.3.0",
]

[project.scripts]
server = "blindspot_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["blindspot_env", "blindspot_env.server"]
package-dir = { "blindspot_env" = ".", "blindspot_env.server" = "server" }
